removed the background generator part of the program: spinning it off for its own stand alone program
